# Conflicts:
#	src/com/company/ui/MainApplication.java

+++ src/com/company/domain/Exam.java
@@ -8,7 +8,7 @@ public class Exam {
     private int duration;
 
     //
-    private short year;
+    private int year;
     private int sem;
     private String tea_id;
     private String c_id;
@@ -60,7 +60,7 @@ public class Exam {
     }
 
     public Exam(String testNo, Timestamp start, int duration,
-                short year, int sem, String tea_id, String c_id, String sub_id){
+                int year, int sem, String tea_id, String c_id, String sub_id){
         this.testNo = testNo;
         this.start = start;
         this.duration = duration;
